Profile pic delete updated

+++ routes/postRoute.js
@@ -5,13 +5,18 @@ const User = require("../models/userModel");
 const Post = require("../models/postModel");
 
 const authorizeUser = require("../middlewares/authorizeUser");
+const { getStorage, ref, deleteObject, } = require("firebase/storage");
+const { initializeApp } = require("firebase/app");
+const firebaseConfig = require("../firebase/firebaseConfig")
+const app = initializeApp(firebaseConfig);
+const storage = getStorage(app);
 
 router.post("/createNewPost", authorizeUser, async (req, res) => {
     try {
         const userId = req.userId;
         const { postType, postUrl, postTitle, postDescription, timeStamp, docGivenName } = req.body;
         if (!postType || !postUrl || !postTitle || !timeStamp || !docGivenName) {
-            res.status(400).json({
+            return res.status(400).json({
                 success: false,
                 message: "Invalid credentials # There must be postType, postTitle, postUrl, timestamp, docGivenName in req.body"
             })
@@ -29,7 +34,7 @@ router.post("/createNewPost", authorizeUser, async (req, res) => {
         })
         await newPost.save();
 
-        res.json({
+        return res.json({
             success: true,
             message: "New post created.",
             data: {
@@ -39,7 +44,7 @@ router.post("/createNewPost", authorizeUser, async (req, res) => {
 
     } catch (error) {
         console.log("Some error in creating new post : ", error);
-        res.status(500).json({
+        return res.status(500).json({
             success: false,
             message: "Internal server error",
         });
@@ -56,17 +61,16 @@ router.get("/fetchPosts", authorizeUser, async (req, res) => {
                 message: "Not post type provided in query"
             })
         }
-        const findPosts = await Post.find({postType:"gallery"})
+        const findPosts = await Post.find({ postType: "gallery" })
 
         return res.json({
-            success:true,
+            success: true,
             message: `All ${postType} posts sent`,
-            data : {
-                postsLength : findPosts.length,
-                posts : findPosts.reverse()
-            }        
+            data: {
+                postsLength: findPosts.length,
+                posts: findPosts.reverse()
+            }
         })
-
     } catch (error) {
         console.log("Some error in fetching post : ", error);
         return res.status(500).json({
@@ -77,31 +81,44 @@ router.get("/fetchPosts", authorizeUser, async (req, res) => {
 })
 
 
-router.delete("/deletePost", authorizeUser, async (req,res)=>{
+router.delete("/deletePost", authorizeUser, async (req, res) => {
     try {
-       const userId = req.userId;
-       const {postId} = req.query;
+        const userId = req.userId;
+        const { postId, postYear, givenName } = req.query;
+        const findPost = await Post.findById(postId);
+        const findUser = await User.findById(userId);
 
-       const findUser = await User.findById(userId);
-       if (findUser.isSpecialUser !== "admin") {
-         return res.status(400).json({
-            success : false,
-            message : "Not authorized to perform task # User can't perform this delete task"
-         })
-       }
+        if (findUser.isSpecialUser !== "admin") {
+            return res.status(400).json({
+                success: false,
+                message: "Not authorized to perform task # User can't perform this delete task"
+            })
+        }
 
-       if (!postId) {
-          return res.status(401).json({
-            success : false,
-            message : "No post id provided # there should be a postId to delete post"
-          })
-       }
+        if (!postId || !postYear || !givenName) {
+            return res.status(401).json({
+                success: false,
+                message: " Invalid credentials to delete a post # there should be a No post id , post year , post name provided to delete post"
+            })
+        }
 
-       await Post.findByIdAndDelete(postId)
-        return res.json({
-            success:true,
-            message : "Post deleted! successfully"
-        })
+        // --- delete post in firebase ----   
+        const deletePostRef = ref(storage, `images/gallery/${postYear}/${findPost.postDetails.docGivenName}`)
+        deleteObject(deletePostRef)
+            .then(async () => {
+                await Post.findByIdAndDelete(postId)
+                return res.json({
+                    success: true,
+                    message: "Post deleted successfully!"
+                })
+            })
+            .catch((error) => {
+                console.log("Error in deleting post", error);
+                return res.status(401).json({
+                    success: false,
+                    message: "Post delete failed!"
+                })
+            });
     } catch (error) {
         console.log("Some error in deleting post : ", error);
         return res.status(500).json({
@@ -112,3 +129,11 @@ router.delete("/deletePost", authorizeUser, async (req,res)=>{
 })
 
 module.exports = router;
+
+/*
+
+Group_Admin_Tue Nov 21 2023 17:54:47 GMT+0530 (India Standard Time)_g6afuD
+Group_Admin_Tue Nov 21 2023 17:54:47 GMT 0530 (India Standard Time)_g6afuD
+Group_Admin_Tue Nov 21 2023 17:54:47 GMT+0530 (India Standard Time)_g6afuD
+
+*/
